% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precAppGraph.R
\name{pa_graph}
\alias{pa_graph}
\title{Precautionary approach graph}
\usage{
pa_graph(fish_data, Fpa, Flim, Bpa, Blim)
}
\arguments{
\item{fish_data}{data.frame representing fishing information for each years.
It contains at least : \itemize{
  \item year
  \item SSB : spawning stock biomass
  \item F : Fishing mortality
}}

\item{Fpa}{integer. Preacautionary approach fishing mortality}

\item{Flim}{integer. Limit fishing mortality}

\item{Bpa}{integer. Preacautionary approach Biomass of the spawning part}

\item{Blim}{integer. Limit Biomass of the spawning part}
}
\value{
The function return a ggplot object.
}
\description{
Precautionary approach graph
}
